Remove entry point, move script to main package
The entry point is not working properly when rrc_2022_datasets is
installed system-wide but the policy that is passed to it is in a
package that is installed to a venv.

To avoid this issue, remove the entry script and instead rely on
`python3 -m  ...`.  To make the command shorter, move
`evaluate_pre_stage.py` out of the `scripts` subpackage to the main
package.  So now it can be called with

    python3 -m rrc_2022_datasets.evaluate_pre_stage ...